third_party/blink/web_tests/transitions/transition-inherit-initial-unprefixed-expected.txt

Test that inherit and initial works on unprefixed transitions.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing inherit.
PASS computedStyle.transitionProperty is 'width'
PASS computedStyle.webkitTransitionProperty is 'width'
PASS computedStyle.transitionDuration is '5s'
PASS computedStyle.webkitTransitionDuration is '5s'
PASS computedStyle.transitionTimingFunction is 'linear'
PASS computedStyle.webkitTransitionTimingFunction is 'linear'
PASS computedStyle.transitionDelay is '2s'
PASS computedStyle.webkitTransitionDelay is '2s'
Testing initial.
PASS computedStyle.transitionProperty is 'all'
PASS computedStyle.webkitTransitionProperty is 'all'
PASS computedStyle.transitionDuration is '0s'
PASS computedStyle.webkitTransitionDuration is '0s'
PASS computedStyle.transitionTimingFunction is 'ease'
PASS computedStyle.webkitTransitionTimingFunction is 'ease'
PASS computedStyle.transitionDelay is '0s'
PASS computedStyle.webkitTransitionDelay is '0s'
PASS successfullyParsed is true

TEST COMPLETE
